Adds delete_extracted_comments.py script
This script removes extracted comment from pot files. In the unknown file outputted from SmartGit, an activity log serving as a clue to understand where the message was displayed in the GUI is outputted as "extracted comment". This information is also useful during the editing of pot files, so it will be maintained until explicitly deleted by this script.